shopur/requirements.txt
Django==5.2.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
django-prometheus==2.3.1
python-dotenv==1.0.1
cryptography==46.0.3
psycopg2-binary==2.9.9
gunicorn==21.2.0

# Required by orders PDF generation
reportlab
